Add required permission to service account
Given that we're using service account, which is required to get
OAuth tokens needed for Code Assist, we have switched from
"Direct Workload Identity Federation" to "Workload Identity Federation through a Service Account".

For this reason, we need to add the right permissions to the
service account.

In future work, we'll remove the direct permissions in WIF.

Co-authored-by: Lee Hagood James  <leehagoodjames@google.com>