feat(actions): use the dispatcher pattern to separate concerns
This introduces a new top-level workflow "gemini-dispatch" that listens for a series of events. Based on the trigger or contents, it calls another workflow (using `workflow_call`). This helps keep each workflow separate and allows us to re-use workflows in the future.

The generic "invoke" command still needs a lot of work, but this is progress.