Auto trigger PR reviews when they are from branches in repo
Pull request events identify members as "contributors", so the
workflows are not auto-triggered. But there are some contributors
who are not members, so workflows would fail on forks.

In this change, we switch the condition for PR events to check for
whether the PR was created from a branch in the repo or from a fork.

Note that only users with write access to the repo can create branches.