feat: support direct wif (#380)
### Why this change?

Fix #379 - Service account is not actually optional, but it says so in
the
[README](https://github.com/google-github-actions/run-gemini-cli#inputs).

### What?

- Add support for direct WIF my making `access_token` and
`gcp_access_token_scopes` into configurable inputs.

### Notes

- This is not a backwards compatibility breaking change, as I added the
current values into the new inputs' defaults.
- Another approach could be to not introduce any new inputs, and instead
have two variants of the authentication step; one that works like today,
and one which doesn't set `access_token` if no service account was
specified. This would require some minor changes to the WIF validation
logic in the action.